requirements.txt
python-dotenv
transformers
torch
Pillow
requests
python-telegram-bot==20.7
nest_asyncio
